refactor(cli): split frontend filter_rules parsing into submodules
